+++ RELEASENOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## Next Release
+
+* Added "fsize" keyword for printing the "friendly size" of files.
+* Fixed a bug that allowed "stats" keyword to be used in the "where" clause.
+
 ## 1.5.0
 
 * Fixed bug where upper case characters in "content contains" searches were not handled properly.